+++ stroom-proxy/stroom-proxy-app/src/main/java/stroom/proxy/app/ProxyOpenIdConfig.java
@@ -12,6 +12,7 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import io.dropwizard.validation.ValidationMethod;
 
 import java.util.List;
+import java.util.Set;
 import javax.validation.constraints.NotNull;
 
 @JsonPropertyOrder(alphabetic = true)
@@ -36,8 +37,9 @@ public class ProxyOpenIdConfig extends AbstractOpenIdConfig implements IsProxyCo
                              @JsonProperty("requestScopes") final List<String> requestScopes,
                              @JsonProperty("clientCredentialsScopes") final List<String> clientCredentialsScopes,
                              @JsonProperty("validateAudience") final boolean validateAudience,
+                             @JsonProperty("validIssuers") final Set<String> validIssuers,
                              @JsonProperty("uniqueIdentityClaim") final String uniqueIdentityClaim,
-        @JsonProperty("userDisplayNameClaim") final String userDisplayNameClaim) {
+                             @JsonProperty("userDisplayNameClaim") final String userDisplayNameClaim) {
         super(identityProviderType,
                 openIdConfigurationEndpoint,
                 issuer,
@@ -52,6 +54,7 @@ public class ProxyOpenIdConfig extends AbstractOpenIdConfig implements IsProxyCo
                 requestScopes,
                 clientCredentialsScopes,
                 validateAudience,
+                validIssuers,
                 uniqueIdentityClaim,
                 userDisplayNameClaim);
     }
@@ -100,6 +103,7 @@ public class ProxyOpenIdConfig extends AbstractOpenIdConfig implements IsProxyCo
                 getRequestScopes(),
                 getClientCredentialsScopes(),
                 isValidateAudience(),
+                getValidIssuers(),
                 getUniqueIdentityClaim(),
                 getUserDisplayNameClaim());
     }