+++ CHANGELOG.md
@@ -12,6 +12,12 @@ DO NOT ADD CHANGES HERE - ADD THEM USING log_change.sh
 ~~~
 
 
+## [v7.1-beta.23-test-open-id-16] - 2023-03-28
+
+* Prevent DB migrations running if stroom is at the desired version.
+
+* Issue **#3112** : Add the `validIssuers` property to allow for non-standard or multiple issuers to be supported.
+
 * Issue **#3305** : Change the formatting of index/data volume limit to show 3 significant figures.
 
 * Issue **#3280** : Make orphan file/meta finder jobs drop out if stroom is shutdown. Change orphan meta finder job to user more efficient SQL. Improve logging output from both finders to include start time, duration and any error/interruption. Fix missing feed names in orphan file finder log.
@@ -4679,7 +4685,8 @@ Issue **gchq/stroom-expression#22** : Add `typeOf(...)` function to dashboard.
 
 * Issue **#202** : Initial release of the new data retention policy functionality.
 
-[Unreleased]: https://github.com/gchq/stroom/compare/v7.1-beta.23...HEAD
+[Unreleased]: https://github.com/gchq/stroom/compare/v7.1-beta.23-test-open-id-16...HEAD
+[v7.1-beta.23-test-open-id-16]: https://github.com/gchq/stroom/compare/v7.1-beta.23-test-open-id-15...v7.1-beta.23-test-open-id-16
 [v7.1-beta.23]: https://github.com/gchq/stroom/compare/v7.1-beta.22...v7.1-beta.23
 [v7.1-beta.22]: https://github.com/gchq/stroom/compare/v7.1-beta.21...v7.1-beta.22
 [v7.1-beta.21]: https://github.com/gchq/stroom/compare/v7.1-beta.20...v7.1-beta.21